refactor(schemas): migrate Phase 17 endpoint schemas — 5 files, 20 classes (#6042)
- audit.py: migrate AuditQueryRequest, AuditQueryResponse, AuditStatisticsResponse,
  CleanupRequest→AuditCleanupRequest → schemas_system.py (imports AuditResult)
- anti_pattern.py: migrate AnalysisRequest→AntiPatternAnalysisRequest, SeveritySummary,
  AntiPatternSummary, AnalysisResponse → schemas_code.py
- analytics_maintenance.py: migrate MaintenanceRecommendationResponse,
  ResourceOptimizationResponse, DashboardResponse, CustomReportRequest → schemas_analytics.py
- analytics_log_patterns.py: migrate LogPattern, LogAnomaly, LogTrend,
  PatternMiningResult → schemas_analytics.py
- long_running_operations.py: migrate CodebaseIndexingRequest, TestSuiteRequest,
  KnowledgeBaseRequest, SecurityScanRequest → schemas_workflows.py (imports PATH)

Renames to avoid generic-name collisions:
- CleanupRequest → AuditCleanupRequest
- AnalysisRequest → AntiPatternAnalysisRequest
