+++ autobot-backend/api/analytics_log_patterns.py
@@ -16,15 +16,18 @@ from typing import Any, Dict, List, Optional, Tuple
 
 import aiofiles
 from fastapi import APIRouter, Depends, HTTPException, Query
-from pydantic import BaseModel, Field
 
 from auth_middleware import check_admin_permission
 from autobot_shared.error_boundaries import ErrorCategory, with_error_handling
 from api.schemas_analytics import (
+    LogAnomaly,
+    LogPattern,
     LogPatternDetailResponse,
     LogPatternHotspotsResponse,
-    LogPatternStatsResponse,
     LogPatternRealtimeResponse,
+    LogPatternStatsResponse,
+    LogTrend,
+    PatternMiningResult,
 )
 
 logger = logging.getLogger(__name__)
@@ -42,58 +45,6 @@ LOG_LEVEL_PRIORITIES = ["CRITICAL", "ERROR", "WARNING", "WARN", "INFO", "DEBUG"]
 # ============================================================================
 # Pydantic Models
 # ============================================================================
-
-
-class LogPattern(BaseModel):
-    """Represents a discovered log pattern"""
-
-    pattern_id: str
-    pattern_template: str
-    occurrences: int
-    first_seen: str
-    last_seen: str
-    log_levels: List[str]
-    sources: List[str]
-    sample_messages: List[str] = Field(default_factory=list, max_length=5)
-    frequency_per_hour: float = 0.0
-    is_error_pattern: bool = False
-    is_anomaly: bool = False
-
-
-class LogAnomaly(BaseModel):
-    """Represents a detected anomaly in logs"""
-
-    anomaly_id: str
-    anomaly_type: str  # spike, gap, new_pattern, error_surge
-    severity: str  # critical, high, medium, low
-    description: str
-    timestamp: str
-    affected_sources: List[str]
-    metric_before: float
-    metric_after: float
-    confidence: float
-
-
-class LogTrend(BaseModel):
-    """Represents a trend in log data"""
-
-    trend_id: str
-    metric_name: str
-    direction: str  # increasing, decreasing, stable
-    change_percent: float
-    time_period: str
-    data_points: List[Dict[str, Any]]
-
-
-class PatternMiningResult(BaseModel):
-    """Result of pattern mining operation"""
-
-    patterns: List[LogPattern]
-    anomalies: List[LogAnomaly]
-    trends: List[LogTrend]
-    summary: Dict[str, Any]
-    analysis_time_ms: float
-    logs_analyzed: int
 
 
 # ============================================================================
